+++ localflavor/generic/validators.py
@@ -15,15 +15,14 @@ from .countries.iso_3166 import ISO_3166_1_ALPHA2_COUNTRY_CODES
 #
 # https://www.swift.com/standards/data-standards/iban
 #
-# The IBAN_COUNTRY_CODE_LENGTH dictionary has been updated version 94 of the IBAN Registry document which was published
-# in April 2023.
+# The IBAN_COUNTRY_CODE_LENGTH dictionary has been updated version 101 of the IBAN Registry document which was published
+# in December 2025.
 #
 # Other Resources:
 #
 # https://en.wikipedia.org/wiki/International_Bank_Account_Number#IBAN_formats_by_country
 # http://www.ecbs.org/iban/france-bank-account-number.html
-# https://www.nordea.com/V%C3%A5ra+tj%C3%A4nster/Internationella+produkter+och+tj%C3%A4nster/Cash+Management/IBAN+countries/908472.html
-
+# https://www.nordea.com/en/our-services/cash-management/iban-and-bban
 
 IBAN_COUNTRY_CODE_LENGTH = {'AD': 24,  # Andorra
                             'AE': 23,  # United Arab Emirates
@@ -49,6 +48,7 @@ IBAN_COUNTRY_CODE_LENGTH = {'AD': 24,  # Andorra
                             'EG': 29,  # Egypt
                             'ES': 24,  # Spain
                             'FI': 18,  # Finland
+                            'FK': 18,  # Falkland Islands (Malvinas)
                             'FO': 18,  # Faroe Islands
                             'FR': 27,  # France + French Guiana (GF), Guadeloupe (GP), Martinique (MQ), Réunion (RE),
                                        #          French Polynesia (PF), French Southern Territories (TF), Mayotte (YT),
@@ -61,6 +61,7 @@ IBAN_COUNTRY_CODE_LENGTH = {'AD': 24,  # Andorra
                             'GL': 18,  # Greenland
                             'GR': 27,  # Greece
                             'GT': 28,  # Guatemala
+                            'HN': 28,  # Honduras
                             'HR': 21,  # Croatia
                             'HU': 28,  # Hungary
                             'IE': 22,  # Ireland
@@ -81,7 +82,7 @@ IBAN_COUNTRY_CODE_LENGTH = {'AD': 24,  # Andorra
                             'MC': 27,  # Monaco
                             'MD': 24,  # Moldova
                             'ME': 22,  # Montenegro
-                            'MK': 19,  # Macedonia
+                            'MK': 19,  # North Macedonia
                             'MN': 20,  # Mongolia
                             'MR': 27,  # Mauritania
                             'MT': 31,  # Malta
@@ -89,6 +90,7 @@ IBAN_COUNTRY_CODE_LENGTH = {'AD': 24,  # Andorra
                             'NI': 28,  # Nicaragua
                             'NL': 18,  # Netherlands
                             'NO': 15,  # Norway
+                            'OM': 23,  # Oman
                             'PK': 24,  # Pakistan
                             'PL': 28,  # Poland
                             'PS': 29,  # Palestine
@@ -104,6 +106,7 @@ IBAN_COUNTRY_CODE_LENGTH = {'AD': 24,  # Andorra
                             'SI': 19,  # Slovenia
                             'SK': 24,  # Slovakia
                             'SM': 27,  # San Marino
+                            'SO': 23,  # Somalia
                             'ST': 25,  # Sao Tome and Principe
                             'SV': 28,  # El Salvador
                             'TL': 23,  # Timor-Leste
@@ -112,28 +115,40 @@ IBAN_COUNTRY_CODE_LENGTH = {'AD': 24,  # Andorra
                             'UA': 29,  # Ukraine
                             'VA': 22,  # Vatican City State
                             'VG': 24,  # British Virgin Islands
-                            'XK': 20}  # Kosovo (user-assigned country code)
+                            'XK': 20,  # Kosovo (user-assigned country code)
+                            'YE': 30}  # Yemen
 
 
-# Nordea has catalogued IBANs for some additional countries but they are not part of the office IBAN network yet.
+# Nordea has catalogued IBANs for some additional countries but they are not part of the official IBAN network yet.
 #
-# Reference:
-# https://www.nordea.com/V%C3%A5ra+tj%C3%A4nster/Internationella+produkter+och+tj%C3%A4nster/Cash+Management/IBAN+countries/908472.html
+# Last updated from nordea-country-and-currency-list.pdf v2.2 (version is at the bottom of the pdf).
+#
+# References:
+# https://www.nordea.com/en/our-services/cash-management/iban-and-bban
+# https://www.nordea.com/en/doc/nordea-country-and-currency-list.pdf
 
 NORDEA_COUNTRY_CODE_LENGTH = {'AO': 25,  # Angola
+                              'BF': 28,  # Burkina Faso
                               'BJ': 28,  # Benin
-                              'BF': 27,  # Burkina Faso
-                              'CI': 28,  # Ivory Coast
+                              'CF': 27,  # Central African Republic
                               'CG': 27,  # Congo
+                              'CI': 28,  # Ivory Coast
                               'CM': 27,  # Cameroon
-                              'CV': 25,  # Cape Verde
+                              'CV': 25,  # Cabo Verde
                               'DZ': 24,  # Algeria
                               'GA': 27,  # Gabon
+                              'GQ': 27,  # Equatorial Guinea
+                              'GW': 25,  # Guinea-Bissau
                               'IR': 26,  # Iran
+                              'KM': 27,  # Comoros
+                              'MA': 28,  # Morocco
                               'MG': 27,  # Madagascar
                               'ML': 28,  # Mali
                               'MZ': 25,  # Mozambique
-                              'SN': 28}  # Senegal
+                              'NE': 28,  # Niger
+                              'SN': 28,  # Senegal
+                              'TD': 27,  # Chad
+                              'TG': 28}  # Togo
 
 
 @deconstructible