+++ localflavor/kw/utils.py
@@ -3,12 +3,12 @@ from datetime import date
 
 def is_valid_civil_id(cid):
     """
-    Checks the validity of a Kuwaiti Civil ID number
-    by verifying the following:
-      * The number should consist of 12 digits
-      * The first digit should be 1, 2, or 3
-      * The extracted birthdate should be a valid date
-      * The checksum should be equal to the last digit of the Civil ID
+    Checks the validity of a Kuwaiti Civil ID number by verifying the following:
+
+    * The number should consist of 12 digits
+    * The first digit should be 1, 2, or 3
+    * The extracted birthdate should be a valid date
+    * The checksum should be equal to the last digit of the Civil ID
     """
     # Civil ID can only start with 1, 2, or 3 till year 2100
     if len(cid) != 12 or not cid.isdigit() or cid[0] not in ('1', '2', '3'):
@@ -36,7 +36,7 @@ def get_birthdate_from_civil_id(cid):
     by_century = {
         '1': '18',
         '2': '19',
-        '3': '20'
+        '3': '20',
     }
     if cid[0] not in ('1', '2', '3'):
         raise ValueError('Invalid first digit')