+++ tests/settings.py
@@ -7,6 +7,7 @@ DATABASES = {
 
 INSTALLED_APPS = [
     'localflavor',
+    'tests.test_ae',
     'tests.test_au',
     'tests.test_br',
     'tests.test_by',
@@ -22,6 +23,7 @@ INSTALLED_APPS = [
     'tests.test_mx',
     'tests.test_np',
     'tests.test_pk',
+    'tests.test_qa',
     'tests.test_ua',
     'tests.test_us',
 ]
@@ -29,3 +31,11 @@ INSTALLED_APPS = [
 SECRET_KEY = 'spam-spam-spam-spam'
 
 SILENCED_SYSTEM_CHECKS = ('1_7.W001', 'models.W042')
+
+# Django 6.1 changed the blank choice to "- Select an option -" which breaks our tests.
+#
+# USE_BLANK_CHOICE_DASH only works for Django 6.x. For Django >= 7.0 we'll need to set
+# django.db.models.fields.BLANK_CHOICE_LABEL in a test app's ready() method if a global setting for this isn't added.
+#
+# https://github.com/django/django/commit/63c56cda133a85a158502891c40465bc0331d3d9
+USE_BLANK_CHOICE_DASH = True